Move tests to correct files
